update: add IntroRoom
